log limitation for single file (core.log) and for directory with file schema
