+++ CommonLogic/Log/Handlers/limit/FileLimitingLogHandler.php
@@ -3,7 +3,6 @@
 namespace exface\Core\CommonLogic\Log\Handlers\limit;
 
 
-use exface\Core\CommonLogic\Log\Helpers\LogHelper;
 use exface\Core\Interfaces\iCanGenerateDebugWidgets;
 
 /**
@@ -31,12 +30,33 @@ class FileLimitingLogHandler extends LimitingWrapper {
 		$this->filename = $filename;
 		$this->maxDays  = $maxDays;
 
-		$this->filenameFormat = '{filename}-{variable}';
+		$this->filenameFormat = '{filename}-{date}';
 		$this->dateFormat = 'Y-m-d';
 	}
 
 	protected function callLogger(Callable $createLoggerCall, $level, $message, array $context = array(), iCanGenerateDebugWidgets $sender = null) {
-		$createLoggerCall(LogHelper::getFilename($this->filename, $this->dateFormat, $this->filenameFormat))->handle($level, $message, $context, $sender);
+		$createLoggerCall($this->getFilename())->handle($level, $message, $context, $sender);
+	}
+
+	/**
+	 * Returns the log filename with date added to it.
+	 *
+	 * @return string
+	 */
+	protected function getFilename()
+	{
+		$fileInfo = pathinfo($this->filename);
+		$timedFilename = str_replace(
+			array('{filename}', '{date}'),
+			array($fileInfo['filename'], date($this->dateFormat)),
+			$fileInfo['dirname'] . '/' . $this->filenameFormat
+		);
+
+		if (!empty($fileInfo['extension'])) {
+			$timedFilename .= '.'.$fileInfo['extension'];
+		}
+
+		return $timedFilename;
 	}
 
 	/**
@@ -49,7 +69,7 @@ class FileLimitingLogHandler extends LimitingWrapper {
 			return;
 		}
 
-		$logFiles = glob(LogHelper::getPattern($this->filename, $this->filenameFormat));
+		$logFiles = glob($this->getGlobPattern());
 		if ($this->maxDays >= count($logFiles)) {
 			// no files to remove
 			return;
@@ -69,5 +89,20 @@ class FileLimitingLogHandler extends LimitingWrapper {
 				restore_error_handler();
 			}
 		}
+	}
+
+	protected function getGlobPattern()
+	{
+		$fileInfo = pathinfo($this->filename);
+		$glob = str_replace(
+			array('{filename}', '{date}'),
+			array($fileInfo['filename'], '*'),
+			$fileInfo['dirname'] . '/' . $this->filenameFormat
+		);
+		if (!empty($fileInfo['extension'])) {
+			$glob .= '.'.$fileInfo['extension'];
+		}
+
+		return $glob;
 	}
 }
